refactor: 리뷰 사항에 맞게 코드 및 단위 테스트 코드 수정작업
