+++ README.md
@@ -8,20 +8,23 @@
     - [x] 우승자에 대해 계산해주는 로직 필요함.
  3. Main 함수 구현
     - [x] 입력을 받아오고 출력을 진행. 받아온 자동차 갯수만큼의 이름을 구현해야함
- 4. 요구사항 명세에 맞는 리팩토링 진행
-    - [x] 들여쓰기 2 넘지 않도록 구현
-    - [x] 함수 이름 15라인 넘지 않도록 구현
- 5. 단위 테스트 구현
-    - [x] 자동차 random이 구현 범위내에 들어오는지 체크
 
 ### STEP5 리팩토링
 
-
-- [x] 자동차는 이름을 가지고 최초 위치는 0을 가리킨다.
-- [x] 자동차 이름은 5자를 초과할 수 없다.
-- [x] 자동차 이름은 쉼표(,)를 기준으로 구분한다.
 - [x] 0에서 9 사이에서 무작위 값을 구한다.
+- [x] 자동차는 이름을 가지고 최초 위치는 0을 가리킨다
+- [x] 자동차 이름은 5자를 초과할 수 없다.
+- [x] 자동차는 강도 4이상의 엑셀을 밟으면 움직 일 수 있음
+- [x] 자동차는 강도 4미만의 엑셀을 밟으면 움직 일 수 없음
+- [x] 자동차가 한대 이상일 경우 경기를 진행할 수 있다 없는 경우 진행 불가
+- [x] 자동차 이름은 쉼표(,)를 기준으로 구분한다.
 - [x] 전진하는 자동차를 출력할 때 자동차 이름을 같이 출력한다.
 - [x] 자동차 경주 게임을 완료한 후 누가 우승했는지를 알려준다. 우승자는 한 명 이상일 수 있다.
-- [x] 핵심 비지니스 로직을 가지는 객체를 domain패키지, UI관련한 객체를 view패키지에 구현
-- [x] MVC패턴 기반으로 리팩토링, view 패키지의 객체가 domain 패키지에 의존할 수 있음, domain패키지의 객체는 view 패키지에 의존하지 않도록 구현
+
+
+## 프로그램 제약 사항 
+
+- 핵심 비지니스 로직을 가지는 객체를 domain패키지, UI관련한 객체를 view패키지에 구현
+- indent(인덴트, 들여쓰기) depth를 2를 넘지 않도록 구현한다. 1까지만 허용한다.
+- 함수(또는 메소드)의 길이가 15라인을 넘어가지 않도록 구현한다
+- MVC패턴 기반으로 리팩토링, view 패키지의 객체가 domain 패키지에 의존할 수 있음, domain패키지의 객체는 view 패키지에 의존하지 않도록 구현
